refs #5192 introducing a user menu
