+++ plugins/CoreHome/angularjs/menudropdown/menudropdown.html
@@ -1,0 +1,9 @@
+<div piwik-focus-anywhere-but-here="view.showItems=false" class="menuDropdown">
+
+    <span class="title"
+          ng-click="view.showItems=!view.showItems"
+          >{{ title }}</span>
+
+    <div class="items" ng-show="view.showItems" ng-transclude></div>
+
+</div>
