+++ plugins/CoreHome/angularjs/menudropdown/menudropdown.less
@@ -1,0 +1,64 @@
+
+.menuDropdown {
+  display: inline-block;
+  padding-right: 14px;
+  color: @theme-color-link;
+
+  .title {
+    position: relative;
+    cursor: pointer;
+
+    &:after {
+      content: '';
+      position: absolute;
+      top: 5px;
+      right: -15px;
+      color: @theme-color-link;
+      display: inline;
+      content: " \25BC";
+      font-size: 1px;
+      height: 0px;
+      width: 0px;
+      border-left: 4px solid transparent;
+      border-right: 4px solid transparent;
+      border-top: 5px solid #1E93D1;
+    }
+  }
+
+  .items {
+    z-index: 21;
+    position: absolute;
+    max-width: 200px;
+    border: 1px solid @color-silver-l80 !important;
+    background: @theme-color-background-base;
+    max-height: 400px;
+    overflow-y: auto;
+    overflow-x: hidden;
+    padding: 0 !important;
+
+    .item {
+      display: block;
+      color: @theme-color-text !important;
+      text-decoration: none !important;
+      padding: 1px 25px 1px 3px !important;
+
+      &:hover {
+        background: @color-silver-l80;
+      }
+
+      &.active {
+        background-color: @color-silver-l80;
+        cursor: default;
+      }
+
+      &.disabled {
+        opacity: 0.5;
+        cursor: default;
+
+        &:hover {
+          background: @theme-color-background-base;
+        }
+      }
+    }
+  }
+}
