+++ plugins/CoreHome/templates/_topBarHelloMenu.twig
@@ -1,20 +1,55 @@
 <div id="topRightBar">
     {% set helloAlias %}
         {% if userAlias is not empty %}
-            <strong>{{ userAlias }}</strong>
+            {{ userAlias }}
         {% else %}
-            <strong>{{ userLogin }}</strong>
+            {{ userLogin }}
         {% endif %}
     {% endset %}
-    <span class="topBarElem">{{ 'General_HelloUser'|translate(helloAlias|trim)|raw }}</span>
-    {% if userLogin != 'anonymous' %}
-        |
-        {% if isAdminLayout is defined %}
-            <span class="topBarElem topBarElemActive">{{ 'General_Settings'|translate }}</span>
+
+    {% macro userMenuItem(label, menu, currentModule, currentAction) %}
+        {% if (menu._url.module == currentModule and (menu._url.action is empty or menu._url.action == currentAction)) %}
+            <a class="active item"
+               {% if menu._tooltip is defined %}title="{{ menu._tooltip }}"{% endif %}
+                    >{{ label|translate }}</a>
         {% else %}
-            <span class="topBarElem"><a href='index.php?module=CoreAdminHome'>{{ 'General_Settings'|translate }}</a></span>
+            <a class="item"
+               href="index.php{{ menu._url|urlRewriteWithParameters }}"
+               {% if menu._tooltip is defined %}title="{{ menu._tooltip }}"{% endif %}
+                    >{{ label|translate }}</a>
         {% endif %}
-    {% endif %}
+    {% endmacro %}
+
+    <span class="topBarElem">
+        <div title="{{ helloAlias|trim|raw }}"
+             piwik-menudropdown>
+
+            {% if userLogin != 'anonymous' %}
+                {% if isAdminLayout is defined %}
+                    <a class="item active">{{ 'General_Settings'|translate }}</a>
+                {% else %}
+                    <a class="item" href="index.php?module=CoreAdminHome">{{ 'General_Settings'|translate }}</a>
+                {% endif %}
+            {% endif %}
+
+            {% for lev1UserLabel,lev1UserMenu in userMenu if lev1UserLabel|slice(0,1) != '_' %}
+
+                {% if userLogin != 'anonymous' or not loop.first %}
+                    <span class="item disabled">--------</span>
+                {% endif %}
+
+                {% if lev1UserMenu._hasSubmenu is defined and lev1UserMenu._hasSubmenu %}
+                    {% for lev2Label,lev2Menu in lev1UserMenu if lev2Label|slice(0,1) != '_' %}
+                        {{ _self.userMenuItem(lev2Label, lev2Menu, currentModule, currentAction) }}
+                    {% endfor %}
+                {% else %}
+                    {{ _self.userMenuItem(lev1UserLabel, lev1UserMenu, currentModule, currentAction) }}
+                {% endif %}
+
+            {% endfor %}
+        </div>
+    </span>
+
     | <span class="topBarElem">
         {% if userLogin == 'anonymous' %}
             <a href='index.php?module={{ loginModule }}'>{{ 'Login_LogIn'|translate }}</a>
