+++ plugins/DBStats/Menu.php
@@ -8,15 +8,8 @@
  */
 namespace Piwik\Plugins\DBStats;
 
-use Exception;
-use Piwik\Common;
-use Piwik\Db;
-use Piwik\Menu\MenuAbstract;
 use Piwik\Menu\MenuAdmin;
-use Piwik\Menu\MenuReporting;
-use Piwik\Menu\MenuTop;
 use Piwik\Piwik;
-use Piwik\Site;
 
 /**
  */
