+++ plugins/DevicesDetection/DevicesDetection.php
@@ -15,9 +15,6 @@ use Piwik\ArchiveProcessor;
 use Piwik\CacheFile;
 use Piwik\Common;
 use Piwik\Db;
-use Piwik\Menu\MenuAbstract;
-use Piwik\Menu\MenuAdmin;
-use Piwik\Menu\MenuMain;
 use Piwik\Piwik;
 use Piwik\Plugin\ViewDataTable;
 use Piwik\WidgetsList;
