+++ plugins/Widgetize/Menu.php
@@ -1,0 +1,24 @@
+<?php
+/**
+ * Piwik - Open source web analytics
+ *
+ * @link http://piwik.org
+ * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
+ *
+ */
+namespace Piwik\Plugins\Widgetize;
+
+use Piwik\Menu\MenuUser;
+use Piwik\Piwik;
+
+class Menu extends \Piwik\Plugin\Menu
+{
+    public function configureUserMenu(MenuUser $menu)
+    {
+        $tooltip   = Piwik::translate('Widgetize_TopLinkTooltip');
+        $urlParams = array('module' => 'Widgetize', 'action' => 'index', 'segment' => false);
+
+        $menu->add('CorePluginsAdmin_MenuPlatform', 'General_Widgets', $urlParams, true, 5, $tooltip);
+    }
+
+}
