+++ plugins/Widgetize/Widgetize.php
@@ -25,17 +25,8 @@ class Widgetize extends \Piwik\Plugin
         return array(
             'AssetManager.getJavaScriptFiles'        => 'getJsFiles',
             'AssetManager.getStylesheetFiles'        => 'getStylesheetFiles',
-            'Menu.Top.addItems'                      => 'addTopMenu',
             'Translate.getClientSideTranslationKeys' => 'getClientSideTranslationKeys'
         );
-    }
-
-    public function addTopMenu(MenuTop $menu)
-    {
-        $tooltip   = Piwik::translate('Widgetize_TopLinkTooltip');
-        $urlParams = array('module' => 'Widgetize', 'action' => 'index', 'segment' => false);
-
-        $menu->add('General_Widgets', null, $urlParams, true, 5, $tooltip);
     }
 
     public function getJsFiles(&$jsFiles)
