+++ plugins/Zeitgeist/stylesheets/ui/_header.less
@@ -60,7 +60,7 @@
 }
 
 #topLeftBar {
-  margin-right: 40px;
+  margin-right: 5px;
 }
 
 .topBarElem {
