+++ plugins/Zeitgeist/stylesheets/ui/_languageSelect.less
@@ -7,6 +7,10 @@
   border: 1px solid #1F447F!important;
 }
 
+#languageSelection {
+  padding-right: 11px;
+}
+
 #languageSelection > form {
   display: inline-block;
 }
